Extract ComposerViewModel to its own file
